Cleaned up the code and adjusted the directory file structure.
