+++ src/CatLib.Core.Tests/IO/TestsStreamExtension.cs
@@ -9,12 +9,12 @@
  * Document: https://catlib.io/
  */
 
-using CatLib.Support;
+using CatLib.IO;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.IO;
 using System.Text;
 
-namespace CatLib.Tests.Support
+namespace CatLib.Tests.IO
 {
     [TestClass]
     public class TestsStreamExtension
