+++ src/CatLib.Core.Tests/Util/TestsArr.cs
@@ -9,14 +9,14 @@
  * Document: https://catlib.io/
  */
 
-using CatLib.Support;
+using CatLib.Util;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
 
 #pragma warning disable CA1031
 
-namespace CatLib.Tests.Support
+namespace CatLib.Tests.Util
 {
     [TestClass]
     public sealed class TestsArr
