+++ src/CatLib.Core/EventDispatcher/EventDispatcher.cs
@@ -10,7 +10,7 @@
  */
 
 using CatLib.Exception;
-using CatLib.Support;
+using CatLib.Util;
 using System;
 using System.Collections.Generic;
 using System.Linq;
